Fixed Tracker 2695089 - File Size overflowed
git-svn-id: https://videomonkey.svn.sourceforge.net/svnroot/videomonkey/trunk@190 ae59dd93-ecb8-44ce-936a-9b63203e4a5a

+++ classes/Transcoder.h
@@ -119,7 +119,7 @@ typedef enum FileStatus { FS_INVALID, FS_VALID, FS_ENCODING, FS_PAUSED, FS_FAILE
 -(BOOL) hasInputAudio;
 -(NSString*) inputVideoFormat;
 -(NSString*) outputFileName;
--(int) outputFileSize;
+-(double) outputFileSize;
 -(NSString*) tempAudioFileName;
 -(NSString*) passLogFileName;
